backend/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cvzone==1.5.6
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python==4.7.0.72
pymongo==4.3.3
python-decouple==3.8
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.2
Werkzeug==2.3.4
